ShaderMeowMeow : The beginning !
So I was asked if I could help porting ShaderMotion to Godot.
Given the beast effort this requires, and the sheer number
of pitfalls this leads to, I decided to first have a little
analyzer that describes the decoding motion STEP by STEP.

That way anyone can clearly identify when I b0rked a step
and tell me what's wrong.

Signed-off-by: Voyage <voyage@miouyouyou.fr>