Midway commit : Adding Mecanim information
I started to add the information provided by Mecanim and
reencoded by Lox on its Javascript version.

I still have to check the values there, since, the Javascript
version might have been modified to use OpenGL orientation.
Mixing this with data that is focusing on Unity orientation
might lead to bogus data in the end.

Still, most of the ShaderMotion ports just contempt themselves
in copying the JS and CS files as-is.
Lox write optimized code, but that's **really** hard to read
and, when doing ports, you want the maximum amount of
readability to understand what went wrong when the project
does.

So here, the data structures focus on readability, with
an inane amount of verbosity.
Verbosity can be trimmed down later, though.

Signed-off-by: Voyage <voyage@miouyouyou.fr>