Added SwingTwist !
Which output the same values as Unity.

Took me a lot of time to put the test together between
Unity and Godot, and making sure that both compute data
the same way.
However, the first basic tests seems to output the exact
same result, so we're good for that part.

Fuck you Unity for the extreme rounding in Debug.Log,
when outputing float values ! I was wondering why
Godot result was like (0.25????, 0, 0, 0,96????) and
why Unity was like (0.3, 0, 0, 1).

And fuck you Godot for not specifying in which units
the angles are when using Quaternion(axis, angle) !
Unity DOES specify that the angles are in degrees.
Godot will use radians instead. I guess that Godot
use radians everywhere, but still, specifying the
unit avoids spending time trying to figure out which
unit the system uses.
Don't expect everyone to be a master in your engine
of reading every documentation page you generate,
even though you'd hope to.

Signed-off-by: Voyage <voyage@miouyouyou.fr>