Starting to add set_hips_position_rotation
But I really need to test this method on Unity and compare
the values with Godot.
The whole idea is still to have the same values as Unity and
do space conversions at the end.

For that I'll have to first implement the HumanPoser constructor,
add some tests and then start testing set_hips_position_rotation.

Implementing the HumanPoser will also need to the addition of
set_bone_swing_twists, so this will be next, after the constructor.

Signed-off-by: Voyage <voyage@miouyouyou.fr>