Starting the work towards automatic testing
And, for the moment, after checking "manually" the result
of the first "swingtwist", the math seems correct and
provide the same values as Unity.

So I made a Dumper in Unity, in order to dump some state
and input/output data when calling methods. The data
are dumped into a JSON file which is now parsed from Godot.

The parsing is still rudimentary at that point, but we're
getting there.
That way, I know which data are used during the execution,
and, that way, I can check things on the Godot side, without
adding tons of Debug.LogError in Unity and printerr in Godot.

Signed-off-by: Voyage <voyage@miouyouyou.fr>