Midway commit : Reimplementing the shader in GDScript
Since I'm not sure that my gray decoder is correct,
I'm actually using RenderDoc to step through the
ShaderMotion shader in a Unity Scene, and reimplement
each step in GDScript while checking that the algorithms
are sending back the same values.

This of course lead to some issues, texture sampling
from the CPU with Godot 4 is a pain when the Image is
compressed for a reason or another.

But the most painful is reading horrible HLSL one-liners.
HLSL is **veeeerry** lenient in its syntax and allows
for extremely hard to read code, especially when it comes
to conditional evaluation, where the condition works
"per-component" and you have no idea about the priority
of evaluation of the various parts, like :

  // Example of interview worthy HLSL one-liner
  a.xz = bool_value && round(x) == float2(0, 1) ? b.zw : 1-round(x);

I'll let you meditate on that one.
Hint : Conditions evaluation work PER-COMPONENT.
Fuck HLSL for allowing that kind of crap.

Signed-off-by: Voyage <voyage@miouyouyou.fr>