Getting there...
The values output by _sm_gray_decoder_sum appear correct...
But I only tested with one set of values, so now I'll have to
test this with all values, if possible.

That said... we're getting there. If all the values appear
correct, all I have to do is to use these gray coding algorithms
with the current slot analyzer and we should be good to go.

The SwingTwist is also correct, so once the adaption done, I'll
see what's still missing.

Signed-off-by: Voyage <voyage@miouyouyou.fr>