Getting the right values now
Roughly, though, since I'm not sampling the colors
at the exact same locations, it seems.
But the values are very close, so I consider that
Gray decoding and SwingTwist are now correct for
the moment.

Now, the next step will be to decode the rotation
correctly and have the transforms being correct,
especially the hips, which decoding code isn't
added yet.

Signed-off-by: Voyage <voyage@miouyouyou.fr>