Now with (invalid) data exporter !
The main scene has changed to a scene that analyze
ShaderMotion tiles based on their Mecanim Bones counterpart.

However, Hips data are still not computed correctly.

That said, the main scene now provide a TSV exporter and
actually parse the pixels from a PNG on the fly, meaning
that I should be able to adapt it to read other video frame
PNG file and analyze the content on the fly.

The reorganization should make it easier to compute Hips
data actually, so this will be the next step.

Signed-off-by: Voyage <voyage@miouyouyou.fr>