Found it ! Now with correct Hips values !
Good thing that I'm printing the values of almost
each step on the screen.
That helped me quickly pinpoint the issue.

Turns out that ShaderMotion only multiply SwingTwist
values by 180. However, the Hips Position is not a
SwingTwist value, it's the actual data themselves,
so no need to multiply.

Alright, next step will be the ApplyX steps !

Signed-off-by: Voyage <voyage@miouyouyou.fr>