Starting to implement apply_transform
This part still require a lot of preparation, and I still
have no idea how these data could be prepared on Godot.

Still, we're getting there, the algorithms seem to provide
the right values when compared to Unity, so the last step
will be to mimic skeleton values as read on Unity, feed
the skeleton and motion data to the method and visualize
the results.

If all the results appear correct, we'll move to the final
part : Make the values useable on Godot.

Signed-off-by: Voyage <voyage@miouyouyou.fr>