Refactored the code to speed up the export process
And also to make things a bit more clean, and let each
object only handle their own things.

So now ParseMotions have a new method named "generate_from"
which take some "tiles data" and a Callable that will be
called with these "tiles data" and a ShaderMotion slot index,
in order to the colors of the tiles associated with this
slot.
It's up to the passed method to understand how to get the
colors from the tiles data.

That way, the "tiles data" format can change at any time,
as long as the provided method knows how to handle it.

Which means that, now, no scene is needed to actually
parse the data, which speed up the parsing process tremendously,
to a point where the whole processing is done in roughly
one second on my system.

A lot of Unity backported methods have been moved to UnityHelpers.
Some other methods were ported to GodotHelpers.

The code should look a bit more clean here and there,
and gained some flexilibility. However, the scenes used
to analyze and display data were not updated, so they might
be unuseable in their current state.

Signed-off-by: Voyage <voyage@miouyouyou.fr>